chore(docker.yml): update Docker workflow to build and deploy to Docker Hub
The changes in this commit update the Docker workflow to build and deploy the application to Docker Hub. The previous workflow was outdated and had unnecessary steps. The new workflow performs the following tasks:

- Only triggers the workflow when pushing to the 'digital-democracy-bot' branch.
- Sets up QEMU and Docker Buildx for multi-platform builds.
- Logs in to Docker Hub using the provided credentials.
- Builds and pushes the Docker image for multiple platforms (linux/amd64, linux/arm64, linux/arm/v7).
- Sets the tags for the Docker image to 'communityfirst/support-whatsapp-bot:latest'.
- Updates the Docker Hub description with a short description of the application.

These changes improve the Docker workflow by simplifying it and ensuring that the Docker image is built and deployed correctly to Docker Hub.